Use double precision

+++ src/binary_skeleton/binary_length.cpp
@@ -6,9 +6,9 @@
 void binary_converter::add_feature(
     const std::string& key,
     const std::string& value, 
-    std::vector<std::pair<std::string, float> >& ret_fv) const {
+    std::vector<std::pair<std::string, double> >& ret_fv) const {
   std::cout << "length =" << value.length() << std::endl;
-  ret_fv.push_back(make_pair(key, (float)  value.length()));
+  ret_fv.push_back(make_pair(key, (double)  value.length()));
 }
 
 extern "C"{
